--- a/reflex/__init__.py
+++ b/reflex/__init__.py
@@ -97,6 +97,7 @@ _ALL_COMPONENTS = [
     "InputLeftElement",
     "InputRightAddon",
     "InputRightElement",
+    "Kbd",
     "Link",
     "LinkBox",
     "LinkOverlay",