--- a/reflex/__init__.py
+++ b/reflex/__init__.py
@@ -308,11 +308,14 @@ def __getattr__(name: str) -> Type:
         # Import the module.
         module = importlib.import_module(_MAPPING[name])
 
+        # Special case for page.
+        if name == "page":
+            globals()["page"] = module.page
+            return module.page
+
         # Get the attribute from the module if the name is not the module itself.
         return (
-            getattr(module, name)
-            if name != _MAPPING[name].rsplit(".")[-1] or name == "page"
-            else module
+            getattr(module, name) if name != _MAPPING[name].rsplit(".")[-1] else module
         )
     except ModuleNotFoundError:
         raise AttributeError(f"module 'reflex' has no attribute {name}") from None