--- a/reflex/components/component.py
+++ b/reflex/components/component.py
@@ -1000,6 +1000,9 @@ def custom_component(
     return wrapper
 
 
+# Alias memo to custom_component.
+memo = custom_component
+
 class NoSSRComponent(Component):
     """A dynamic component that is not rendered on the server."""