--- a/reflex/components/component.py
+++ b/reflex/components/component.py
@@ -1003,6 +1003,7 @@ def custom_component(
 # Alias memo to custom_component.
 memo = custom_component
 
+
 class NoSSRComponent(Component):
     """A dynamic component that is not rendered on the server."""