--- a/reflex/page.py
+++ b/reflex/page.py
@@ -2,8 +2,7 @@
 
 from __future__ import annotations
 
-from reflex.components.component import Component
-from reflex.event import EventHandler
+from typing import Any
 
 DECORATED_PAGES = []
 
@@ -14,8 +13,8 @@ def page(
     image: str | None = None,
     description: str | None = None,
     meta: str | None = None,
-    script_tags: list[Component] | None = None,
-    on_load: EventHandler | list[EventHandler] | None = None,
+    script_tags: list[Any] | None = None,
+    on_load: Any | list[Any] | None = None,
 ):
     """Decorate a function as a page.