@picklelo 2023-11-07
This still needs to work with `.pyi` before merging. Maybe we can generate one for the base `reflex/__init__.py`